Moved cache test adapter to support modules
