+++ test/unit/configuration/matcher_test.exs
@@ -4,27 +4,6 @@ defmodule Annon.Configuration.MatcherTest do
   import ExUnit.CaptureLog
   alias Annon.Configuration.Matcher
 
-  defmodule TestAdapter do
-    @moduledoc false
-    @behaviour Annon.Configuration.CacheAdapter
-    require Logger
-
-    def init do
-      Logger.debug("Adapter initialized")
-      :ok
-    end
-
-    def match_request(_scheme, _method, _host, _port, _path) do
-      Logger.debug("Match called")
-      {:ok, :i_am_api}
-    end
-
-    def config_change do
-      Logger.debug("Config changed")
-      :ok
-    end
-  end
-
   test "raises when adapter is not set" do
     assert_raise RuntimeError, fn ->
       Matcher.init([])
@@ -33,14 +12,14 @@ defmodule Annon.Configuration.MatcherTest do
 
   test "initializes adapter" do
     debug_fn = fn ->
-      Matcher.start_link([adapter: Annon.Configuration.CacheAdapter.ETSTest.TestAdapter])
+      Matcher.start_link([adapter: Annon.CacheTestAdapter])
     end
 
     assert capture_log(debug_fn) =~ "Adapter initialized"
   end
 
   test "uses adapter to match request" do
-    Matcher.start_link([adapter: Annon.Configuration.CacheAdapter.ETSTest.TestAdapter])
+    Matcher.start_link([adapter: Annon.CacheTestAdapter])
 
     debug_fn = fn ->
       assert {:ok, :i_am_api} = Matcher.match_request("https", "POST", "example.com", 80, "/my_path")
@@ -50,7 +29,7 @@ defmodule Annon.Configuration.MatcherTest do
   end
 
   test "notifies adapter on config change" do
-    Matcher.start_link([adapter: Annon.Configuration.CacheAdapter.ETSTest.TestAdapter])
+    Matcher.start_link([adapter: Annon.CacheTestAdapter])
 
     debug_fn = fn ->
       Matcher.config_change()
